requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
tweepy>=4.14.0
Pillow>=10.0.0
anthropic>=0.40.0
openai>=1.52.0
feedparser>=6.0.0
requests>=2.31.0
requests-oauthlib>=2.0.0
google-api-python-client>=2.149.0
google-auth>=2.35.0
